Mohammadnezhad: Implement Publisher functionality 
Mohammadnezhad: Implement Gender functionality

+++ ParsaLibraryManagementProject/Libraries/ParsaLibraryManagement.Domain/Entities/Publisher.cs
@@ -23,7 +23,7 @@
         /// <summary>
         /// Gets or sets the gender ID of the publisher.
         /// </summary>
-        public byte GenderId { get; set; }
+        public short GenderId { get; set; }
 
         /// <summary>
         /// Gets or sets the email of the publisher.
